Add show_linetype and show_shape
Add `show_linetype()` and `show_shape()`, which are analagous to `show_col()` for line types and shape palettes, respectively.

These functions are currently in the **ggthemes** package, but they would seem to be general enough and useful enough to belong in the **scales** package.

Changes as part of this update:

-  moved source code location of `show_col()` so all `show_*()` are in the same file.
-  added examples to `linetype_pal` and `linetype_shape`
-  updated `show_col()` to invisibly return its input, like `show_linetype` and `show_shape`.